Style TxReviewPage
