+++ lib/constants/strings.dart
@@ -187,8 +187,8 @@ These words need to be written down carefully and in the exact order shown.''';
   static const String assetShortNameHint = 'YLW JACK2, max 8 characters.';
   static const String amountHint = '200000';
   static const String totalTxFee = 'Total Transaction Fee';
-  static const String cancel = 'CANCEL';
-  static const String confirm = 'CONFIRM';
+  static const String cancel = 'Cancel';
+  static const String confirm = 'Confirm';
   static const String sending = 'Sending';
   static const String from = 'From';
   static const String to = 'To';
