+++ lib/presentation/transfers/asset_transfer_input_page.dart
@@ -4,14 +4,16 @@ import 'package:flutter_svg/flutter_svg.dart';
 import 'package:ribn/constants/assets.dart';
 import 'package:ribn/constants/strings.dart';
 import 'package:ribn/containers/asset_transfer_input_container.dart';
-import 'package:ribn/presentation/transfers/bottom_review_button.dart';
+import 'package:ribn/presentation/transfers/bottom_review_action.dart';
 import 'package:ribn/presentation/transfers/transfer_utils.dart';
 import 'package:ribn/presentation/transfers/widgets/from_address_field.dart';
 import 'package:ribn/utils.dart';
 import 'package:ribn/widgets/address_display_container.dart';
+import 'package:ribn/widgets/fee_info.dart';
 import 'package:ribn_toolkit/constants/colors.dart';
 import 'package:ribn_toolkit/constants/styles.dart';
 import 'package:ribn_toolkit/widgets/atoms/custom_page_title.dart';
+import 'package:ribn_toolkit/widgets/atoms/large_button.dart';
 import 'package:ribn_toolkit/widgets/molecules/asset_amount_field.dart';
 import 'package:ribn_toolkit/widgets/molecules/asset_selection_field.dart';
 import 'package:ribn_toolkit/widgets/molecules/loading_spinner.dart';
@@ -74,9 +76,6 @@ class _AssetTransferInputPageState extends State<AssetTransferInputPage> {
 
   @override
   Widget build(BuildContext context) {
-    final bool enteredValidInputs =
-        _validRecipientAddress.isNotEmpty && _amountController.text.isNotEmpty && _validAmount;
-
     return AssetTransferInputContainer(
       builder: (BuildContext context, AssetTransferInputViewModel vm) {
         return Scaffold(
@@ -236,34 +235,61 @@ class _AssetTransferInputPageState extends State<AssetTransferInputPage> {
               _loadingRawTx ? const LoadingSpinner() : const SizedBox(),
             ],
           ),
-          bottomNavigationBar: BottomReviewButton(
-            vm: vm,
-            onPressed: enteredValidInputs
-                ? () {
-                    setState(() {
-                      _loadingRawTx = true;
-                    });
-                    vm.initiateTx(
-                      recipient: _validRecipientAddress,
-                      amount: _amountController.text,
-                      note: _noteController.text,
-                      assetCode: _selectedAsset!.assetCode,
-                      assetDetails: vm.assetDetails[_selectedAsset!.assetCode.toString()],
-                      onRawTxCreated: (bool success) async {
-                        _loadingRawTx = false;
-                        setState(() {});
-                        // Display error dialog if failed to create raw tx
-                        if (!success) {
-                          await TransferUtils.showErrorDialog(context);
-                        }
-                      },
-                    );
-                  }
-                : () {},
-            enteredValidInputs: enteredValidInputs,
+          bottomNavigationBar: BottomReviewAction(
+            children: Column(
+              crossAxisAlignment: CrossAxisAlignment.start,
+              children: [
+                // fee info for the tx
+                FeeInfo(fee: vm.networkFee),
+                _buildReviewButton(vm),
+              ],
+            ),
           ),
         );
       },
+    );
+  }
+
+  Widget _buildReviewButton(AssetTransferInputViewModel vm) {
+    final bool enteredValidInputs =
+        _validRecipientAddress.isNotEmpty && _amountController.text.isNotEmpty && _validAmount;
+
+    return Padding(
+      padding: const EdgeInsets.only(top: 15),
+      child: Center(
+        child: LargeButton(
+          buttonWidth: double.infinity,
+          buttonChild: Text(
+            Strings.review,
+            style: RibnToolkitTextStyles.btnMedium.copyWith(
+              color: Colors.white,
+            ),
+          ),
+          onPressed: enteredValidInputs
+              ? () {
+                  setState(() {
+                    _loadingRawTx = true;
+                  });
+                  vm.initiateTx(
+                    recipient: _validRecipientAddress,
+                    amount: _amountController.text,
+                    note: _noteController.text,
+                    assetCode: _selectedAsset!.assetCode,
+                    assetDetails: vm.assetDetails[_selectedAsset!.assetCode.toString()],
+                    onRawTxCreated: (bool success) async {
+                      _loadingRawTx = false;
+                      setState(() {});
+                      // Display error dialog if failed to create raw tx
+                      if (!success) {
+                        await TransferUtils.showErrorDialog(context);
+                      }
+                    },
+                  );
+                }
+              : () {},
+          disabled: !enteredValidInputs,
+        ),
+      ),
     );
   }
 }
