+++ lib/presentation/transfers/poly_transfer_input_page.dart
@@ -3,16 +3,18 @@ import 'package:flutter_svg/flutter_svg.dart';
 import 'package:ribn/constants/assets.dart';
 import 'package:ribn/constants/strings.dart';
 import 'package:ribn/containers/poly_transfer_input_container.dart';
-import 'package:ribn/presentation/transfers/bottom_review_button.dart';
+import 'package:ribn/presentation/transfers/bottom_review_action.dart';
 import 'package:ribn/presentation/transfers/transfer_utils.dart';
 import 'package:ribn/presentation/transfers/widgets/custom_input_field.dart';
 import 'package:ribn/presentation/transfers/widgets/from_address_field.dart';
 import 'package:ribn/utils.dart';
 import 'package:ribn/widgets/address_display_container.dart';
+import 'package:ribn/widgets/fee_info.dart';
 import 'package:ribn_toolkit/constants/colors.dart';
 import 'package:ribn_toolkit/constants/styles.dart';
 import 'package:ribn_toolkit/widgets/atoms/custom_page_title.dart';
 import 'package:ribn_toolkit/widgets/atoms/custom_text_field.dart';
+import 'package:ribn_toolkit/widgets/atoms/large_button.dart';
 import 'package:ribn_toolkit/widgets/molecules/loading_spinner.dart';
 import 'package:ribn_toolkit/widgets/molecules/note_field.dart';
 import 'package:ribn_toolkit/widgets/molecules/recipient_field.dart';
@@ -75,9 +77,6 @@ class _PolyTransferInputPageState extends State<PolyTransferInputPage> {
 
   @override
   Widget build(BuildContext context) {
-    final bool enteredValidInputs =
-        _validRecipientAddress.isNotEmpty && _amountController.text.isNotEmpty && _validAmount;
-
     return PolyTransferInputContainer(
       builder: (BuildContext context, PolyTransferInputViewModel vm) {
         return Scaffold(
@@ -186,29 +185,15 @@ class _PolyTransferInputPageState extends State<PolyTransferInputPage> {
               _loadingRawTx ? const LoadingSpinner() : const SizedBox(),
             ],
           ),
-          bottomNavigationBar: BottomReviewButton(
-            vm: vm,
-            onPressed: enteredValidInputs
-                ? () {
-                    setState(() {
-                      _loadingRawTx = true;
-                    });
-                    vm.initiateTx(
-                      amount: _amountController.text,
-                      recipient: _validRecipientAddress,
-                      note: _noteController.text,
-                      onRawTxCreated: (bool success) async {
-                        _loadingRawTx = false;
-                        setState(() {});
-                        // Display error dialog if failed to create raw tx
-                        if (!success) {
-                          await TransferUtils.showErrorDialog(context);
-                        }
-                      },
-                    );
-                  }
-                : () {},
-            enteredValidInputs: enteredValidInputs,
+          bottomNavigationBar: BottomReviewAction(
+            children: Column(
+              crossAxisAlignment: CrossAxisAlignment.start,
+              children: [
+                // fee info for the tx
+                FeeInfo(fee: vm.networkFee),
+                _buildReviewButton(vm),
+              ],
+            ),
           ),
         );
       },
@@ -259,6 +244,47 @@ class _PolyTransferInputPageState extends State<PolyTransferInputPage> {
             );
           });
         },
+      ),
+    );
+  }
+
+  Widget _buildReviewButton(PolyTransferInputViewModel vm) {
+    final bool enteredValidInputs =
+        _validRecipientAddress.isNotEmpty && _amountController.text.isNotEmpty && _validAmount;
+
+    return Padding(
+      padding: const EdgeInsets.only(top: 15),
+      child: Center(
+        child: LargeButton(
+          buttonWidth: double.infinity,
+          buttonChild: Text(
+            Strings.review,
+            style: RibnToolkitTextStyles.btnMedium.copyWith(
+              color: Colors.white,
+            ),
+          ),
+          onPressed: enteredValidInputs
+              ? () {
+                  setState(() {
+                    _loadingRawTx = true;
+                  });
+                  vm.initiateTx(
+                    amount: _amountController.text,
+                    recipient: _validRecipientAddress,
+                    note: _noteController.text,
+                    onRawTxCreated: (bool success) async {
+                      _loadingRawTx = false;
+                      setState(() {});
+                      // Display error dialog if failed to create raw tx
+                      if (!success) {
+                        await TransferUtils.showErrorDialog(context);
+                      }
+                    },
+                  );
+                }
+              : () {},
+          disabled: !enteredValidInputs,
+        ),
       ),
     );
   }
